+++ packages/manfred-common/src/model/index.ts
@@ -1,2 +1,1 @@
 export * from './manfred-schema.model';
-export * from './manfred-relevant-link.model';
