--- packages/manfred-common/src/model/manfred-relevant-link.model.ts
@@ -1,7 +1,0 @@
-type ManfredRelevantLinkType = 'linkedin' | 'github' | 'twitter' | 'website' | 'other';
-
-export interface ManfredRelevantLink {
-  type: ManfredRelevantLinkType;
-  URL: string;
-  description?: string;
-}
