+++ packages/manfred2md/src/engine/md-parts/profile-section/profile-section.ejs
@@ -13,19 +13,19 @@
   <%_ if (profile?.relevantLinks) { -%>
     <%_ for (const link of profile?.relevantLinks) { -%>
       <%_ if (link.type === 'linkedin') { -%>
-  - ![<% link.type %>](https://raw.githubusercontent.com/Lemoncode/manfred-export-app/main/media/linkedin.svg) [<% link.type %>](<%= link.URL -%>)
-      <%_ } -%>
-      <%_ if (link.type === 'github') { -%>
-  - ![<% link.type %>](https://raw.githubusercontent.com/Lemoncode/manfred-export-app/main/media/github.svg) [<% link.type %>](<%= link.URL -%>)
+  - ![<%= link.type -%>](https://raw.githubusercontent.com/Lemoncode/manfred-export-app/main/media/linkedin.svg) [<%= link.type -%>](<%= link.URL -%>)
       <%_ } -%>
       <%_ if (link.type === 'twitter') { -%>
-  - ![<% link.type %>](https://raw.githubusercontent.com/Lemoncode/manfred-export-app/main/media/twitter.svg) [<% link.type %>](<%= link.URL -%>)
+  - ![<%= link.type -%>](https://raw.githubusercontent.com/Lemoncode/manfred-export-app/main/media/twitter.svg) [<%= link.type -%>](<%= link.URL -%>)
       <%_ } -%>
-      <%_ if (link.type === 'website') { -%>
-  - ![<% link.type %>](https://raw.githubusercontent.com/Lemoncode/manfred-export-app/main/media/website.svg) [<% link.type %>](<%= link.URL -%>)
+      <%_ if (link.type === 'github') { -%>
+  - ![<%= link.type -%>](https://raw.githubusercontent.com/Lemoncode/manfred-export-app/main/media/github.svg) [<%= link.type -%>](<%= link.URL -%>)
       <%_ } -%>
-      <%_ if (link.type === 'other') { -%>
-  - ![<% link.type %>](https://raw.githubusercontent.com/Lemoncode/manfred-export-app/main/media/otros.svg) [<% link.type %>](<%= link.URL -%>)
+      <%_ if (link.type === 'web') { -%>
+  - ![<%= link.type -%>](https://raw.githubusercontent.com/Lemoncode/manfred-export-app/main/media/website.svg) [<%= link.type -%>](<%= link.URL -%>)
+      <%_ } -%>
+      <%_ if (link.type === 'otros') { -%>
+  - ![<%= link.type -%>](https://raw.githubusercontent.com/Lemoncode/manfred-export-app/main/media/otros.svg) [<%= link.type -%>](<%= link.URL -%>)
       <%_ } -%>
     <%_ } %>
   <%_ } -%>
